requirements.txt
pandas>=2.0,<3.0
google-genai
pytest>=8.0,<9.0